local_horizon_developer_rest_api/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-restful
flask-httpauth
flask-cors
psycopg2-binary
boto3
Werkzeug==2.1.1
Flask-Limiter
mysql-connector-python
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
Flask-HTTPAuth
Flask-CORS
warrant
pytest
langchain
openai
pandas
python-dotenv 
requests